requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
youtube-transcript-api
python-dotenv
langchain
langchain-google-genai
langchain-community
faiss-cpu
